Move metrics proto package to stats/v1
In order for projects to link the metrics.proto file with go modules it
is required to have a go import. This creates a problem because the
cgroups package is not platform independent although its types are.
Moving the types to a sperate package allows consumption of the types
regardless of platform.

Signed-off-by: Justin Terry (VM) <juterry@microsoft.com>