test: add basic creation test for systemd
Add a create test for systemd. This currently fails, as test.slice is
not created for each subsystem.

```
=== RUN   TestCreateSystemd
    TestCreateSystemd: cgroup_test.go:80: group hugetlb was not created
--- FAIL: TestCreateSystemd (0.01s)
FAIL
exit status 1
FAIL    github.com/containerd/cgroups   0.011s
```

In non-mock testing, I noticed that deletion was failing, as the slice
still exists on the host after completing the test (and rerunning test
results in a failure). `systemctl stop test.slice` was required to
clean.

Signed-off-by: Eric Ernst <eric@amperecomputing.com>