Retry file writes on EINTR errors
Go 1.14 introduces asynchronous preemption, which results in
applications that invoke slow syscalls (e.g. writing to a cgroup file)
getting EINTR errors. The Go runtime will fix this in version 1.15
onwards:
    https://github.com/golang/go/issues/38033

But this fix won't be backported:
    https://github.com/golang/go/issues/39026#issuecomment-627589315

As such, introduce a 'writeFile' helper function that retries when the
error is an EINTR error.

Signed-off-by: Andrew Dunham <adunham@stripe.com>