v2: Fix inotify leak when cgroup is deleted
When running on cgroup2, we currently leak the inotify instance (and goroutine
blocked on read) used to monitor 'memory.events' on any container exit. This is
highly problematic when containers are automatically restarted because we will
exhaust either the fd limit or system-wide inotify instance limit.

When a process exits, there is no memory event and even when the cgroup is
deleted, the inotify read is also not unblocked. This is not the case when
containerd is running on cgroup (v1) because that uses a different mechanism
for notification and detects cgroup deletion.

Fulfill the contract on cgroup2 by additionally monitoring cgroup.events for
process exit. When the last process exits the kernel signals an event on
'cgroup.events'. For robustness we check both 'cgroup.events' and
'memory.events' on any notification and also handle ENOENT/ENODEV errors from
read/open of 'memory.events'.

We signal exit up the stack by closing the error channel. Strangely, the error
channel was not previously being returned to the caller.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>