make cmd/ a separate module (as it's only for testing)
To my knowledge, the `cmd/cgctl` utility is only for testing, and
its dependencies should therefore not be needed as dependencies for
the main module.

This patch adds a `go.mod` in the `cmd/` directory to make it it's own
module. A replace rule is included in that `go.mod` to make it use the
local version of github.com/containerd/cgroups when building.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>