cgroup2: Add Kill method to manager
This adds in support for killing all of the processes in a cgroup. In
5.14+ this is very simple, a cgroup.kill file exists that all you need
to do is write "1" to https://lwn.net/Articles/855924/. On kernels prior,
or if the file doesn't exist to be more pedantic to account for potential
backports, I've taken the approach runc currently uses which is a manual
process of freezing the cgroup -> sending a signal to all of the processes
-> thawing the cgroup.

This also adds in a simple test for this that should work on 5.15+ and prior
kernels.

Signed-off-by: Danny Canter <danny@dcantah.dev>