cg2: syscall -> x/sys/unix
Change to x/sys/unix definitions. I left cgroups 1
as right now it *should* be able to build on non-linux
platforms, but there's a singular syscall.EINVAL check
which swapping to unix would ruin. cg2 should already not
be able to build on non-linux as we were using InotifyInit
and Statfs_t in tests.

Signed-off-by: Danny Canter <danny@dcantah.dev>