Cgroups2: Reduce allocations for manager.Stat
This change works towards bringing down allocations for manager.Stat().
There's two things this change does:

1. Swap to opening a file and issuing a single read for uint64/"max" values.
Previously we were doing os.ReadFile's which returns a byte slice, so it needs
to allocate.
2. Sizing the map we store {controller}.stat values in. We store 40+ stats in this
map and the default bucket size for Go seems to be smaller than this, so we'd incur
a couple allocs at runtime when adding these.

Signed-off-by: Danny Canter <danny@dcantah.dev>