Cgroup1: Lower buffer size for uint reads
This was partly a brain fart on my end during review for this, we
should only need 20 bytes to handle reading the largest unsigned 64
bit integer (21 if we for some reason wanted to read the line feed
that's in all of these files to). Lets use 32 for a nice power of two
though.

Signed-off-by: Danny Canter <danny@dcantah.dev>