cgroup2: Manager.Delete: handle both "threaded" and "domain threaded"
commit 6f5001dbae8ec248caad9b86c07c869f10ec5726 added special handling
for threaded cgroup types. A later contribution added detection for
"domain threaded" as known type, but did not update the handling to
detect this type.

From the original PR;

> Reading cgroup.procs seems to return ENOTSUPP when threaded, so check
> the type of the cg when going to delete and read the relevant file.

An alternative could be to check both variants unconditionally, and to
error if either Manager.Threads or Manager.Procs is non-zero.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>